Possible 'working' hack: But there is a problem
The problem is that we STILL DO NOT WANT lasso select to be possible
WHILE SOMETHING OTHER THAN THE VecDrawing is selected, I think.

This seems a little annoying, but I think it is the way to go.

The problem is that, say, we want to rotate a Bone2D.. but it doesn't
work, because the lasso select interrupts it.

Now, we could make it so that switching to the Drawing tab automatically
re-selects the drawing. I think this might be somewhat reasonable.